I don’t need gnu-sed after all
To be honest, I’m still not entirely sure what `-z` does in this script.
The option should make sed change its view of what counts as a line.
Normally, sed considers `\n` the end of a line, but after `-z`, sed
allows `\n` within a line and considers a null character the end of
a line.

However, the script works—or appears to work—without the option. Since
macOS doesn’t ship with GNU sed, and I don’t want to install it just for
this, I’m going to see if I can just chuck `-z`.