feat: add a minimal abbreviation plugin
I got the idea from zsh-abbr, but this is my attempt to do very little
as simply as possible.

The plugin does two things. First, if the user enters the key of an
abbreviation followed by a space, the abbreviation key is replaced by
its value. Then the user can continue typing. Second, if the user enters
the key of an abbreviation followed by return, the abbreviation key is
replaced by its value before the line is run.

Abbreviation replacement after a space happens anywhere in a command,
but only the word before return is replaced for execution. This is
because abbreviation is only done interactively. If you paste a command,
abbreviation will not happen. Thus, anything other than the last word of
a line will already have been replaced.

Spaces are not added after replacement. Type your own. I may reconsider
this after using the plugin for a while.